chatbot/pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "A simple chatbot application"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncio>=3.4.3",
    "google-genai>=1.21.1",
    "langchain>=0.3.26",
    "langchain-community>=0.3.26",
    "langchain-google-genai>=2.1.5",
    "langchain-mcp-adapters>=0.1.7",
    "langgraph>=0.4.8",
    "mcp>=1.9.4",
    "nest-asyncio>=1.6.0",
    "python-dotenv>=1.1.0",
    "streamlit>=1.46.0",
]
